feat: fromString
