+++ packages/react/src/experimental/OneCalendar/OneCalendar.tsx
@@ -1,14 +1,15 @@
 import { ChevronLeft, ChevronRight } from "@/icons/app"
 import { Input } from "@/ui/input"
 
-import { useEffect, useMemo, useState } from "react"
+import { useCallback, useEffect, useMemo, useState } from "react"
 import { Button } from "../../components/Actions/Button"
 import { useI18n } from "../../lib/providers/i18n"
 import {
   granularityDefinitions,
   GranularityDefinitionSimple,
 } from "./granularities"
-import { CalendarMode, CalendarView, DateRange } from "./types"
+import { CalendarMode, CalendarView, DateRange, DateRangeString } from "./types"
+import { isValidDate, toDateRange } from "./utils"
 
 export interface OneCalendarProps {
   mode: CalendarMode
@@ -28,7 +29,7 @@ export const getGranularityDefinition = (
     throw new Error(`Granularity definition for view ${view} not found`)
   }
   return {
-    toString: granularity.toString,
+    toRangeString: granularity.toRangeString,
   }
 }
 
@@ -45,21 +46,40 @@ export function OneCalendar({
 
   const [viewDate, setViewDate] = useState<Date>(defaultMonth)
 
-  const [selected, setSelected] = useState<Date | DateRange | null>(
+  const [selected, setSelectedInternal] = useState<Date | DateRange | null>(
     defaultSelected
   )
 
-  useEffect(() => {
-    setSelected(defaultSelected)
-  }, [defaultSelected, setSelected])
-
   const [motionDirection, setMotionDirection] = useState(1)
 
-  const granularity = granularityDefinitions[view]
+  const granularity = useMemo(() => granularityDefinitions[view], [view])
 
-  // Handle navigation
+  const setSelected = useCallback(
+    (date: Date | DateRange | null) => {
+      setSelectedInternal(date)
+
+      // Set the input value
+      setInputValue(granularity.toRangeString(date))
+
+      const newViewDate = granularity.getViewDateFromDate(
+        date instanceof Date ? date : date?.from || date?.to || new Date()
+      )
+
+      if (newViewDate !== granularity.getViewDateFromDate(viewDate)) {
+        setViewDate(newViewDate)
+      }
+    },
+    // eslint-disable-next-line react-hooks/exhaustive-deps -- only needs to be rebuilt when the granularity changes
+    [granularity]
+  )
+
+  useEffect(() => {
+    // setSelected(defaultSelected)
+  }, [defaultSelected, setSelected])
+
+  // Handle ui view navigation
   const navigate = (direction: -1 | 1) => {
-    const newDate = granularity.navigate(viewDate, direction)
+    const newDate = granularity.navigateUIView(viewDate, direction)
     setMotionDirection(direction)
     setViewDate(newDate)
   }
@@ -75,12 +95,10 @@ export function OneCalendar({
     onSelect?.(date)
   }
 
-  const dateString = useMemo(() => {
-    if (!selected) return ["", ""]
-    return granularity.toStringRange(selected)
-  }, [selected, granularity])
-
-  console.log("test")
+  const [inputValue, setInputValue] = useState<DateRangeString>({
+    from: "",
+    to: "",
+  })
 
   const [inputError, setInputError] = useState<{
     from: boolean
@@ -90,17 +108,53 @@ export function OneCalendar({
     to: false,
   })
 
-  const handleInputChange = (input: "from" | "to", value: string) => {
-    const newDate = granularity.fromString([value, value])
-    console.log("newDate", newDate)
-    if (newDate instanceof Error) {
-      setInputError((prev) => ({
-        ...prev,
-        [input]: newDate.message,
-      }))
-      return
+  const handleInputChange = (input: "from" | "to") => {
+    setSelectFromInput(input, inputValue)
+  }
+
+  const setSelectFromInput = (
+    input: "from" | "to",
+    inputValue: DateRangeString
+  ) => {
+    const newDate = granularity.fromString(inputValue)
+    const error = newDate && newDate[input] && !isValidDate(newDate[input])
+    setInputError((prev) => ({
+      ...prev,
+      [input]: error,
+    }))
+
+    if (!error) {
+      setSelected(newDate)
     }
-    setSelected(newDate)
+  }
+  useEffect(() => {
+    const range = toDateRange(selected)
+
+    const { from, to } = granularity.toRangeString(
+      range ? range : { from: new Date(), to: undefined }
+    )
+    setInputValue({
+      from: from || "",
+      to: to || "",
+    })
+  }, [granularity, selected])
+
+  const handleInputNavigate = (input: "from" | "to", direction: -1 | 1) => {
+    const currentDate = inputValue[input]
+      ? granularity.fromString(inputValue[input])
+      : undefined
+    const newDate = currentDate
+      ? granularity.navigate(currentDate.from, direction)
+      : undefined
+
+    if (isValidDate(newDate)) {
+      const newInputValue = {
+        ...inputValue,
+        [input]: granularity.toRangeString(newDate).from,
+      }
+      setInputValue(newInputValue)
+      setSelectFromInput(input, newInputValue)
+    }
   }
 
   return (
@@ -109,16 +163,38 @@ export function OneCalendar({
         <div className="mb-2 flex gap-2">
           <Input
             error={!!inputError.from}
-            value={dateString[0]}
-            onBlur={(e) => handleInputChange("from", e.target.value)}
-            onChange={(e) => console.log("from", e.target.value)}
+            value={inputValue.from}
+            onBlur={() => handleInputChange("from")}
+            onKeyDown={(e) => {
+              if (e.key === "Enter") {
+                handleInputChange("from")
+              }
+              if (e.key === "ArrowUp" || e.key === "ArrowDown") {
+                e.preventDefault()
+                handleInputNavigate("from", e.key === "ArrowDown" ? -1 : 1)
+              }
+            }}
+            onChange={(e) =>
+              setInputValue({ ...inputValue, from: e.target.value })
+            }
           />
           {mode === "range" && (
             <Input
               error={!!inputError.to}
-              value={dateString[1]}
-              onBlur={(e) => handleInputChange("to", e.target.value)}
-              onChange={(e) => console.log("to", e.target.value)}
+              value={inputValue.to}
+              onBlur={() => handleInputChange("to")}
+              onKeyDown={(e) => {
+                if (e.key === "Enter") {
+                  handleInputChange("to")
+                }
+                if (e.key === "ArrowUp" || e.key === "ArrowDown") {
+                  e.preventDefault()
+                  handleInputNavigate("to", e.key === "ArrowDown" ? -1 : 1)
+                }
+              }}
+              onChange={(e) =>
+                setInputValue({ ...inputValue, to: e.target.value })
+              }
             />
           )}
         </div>
