+++ packages/react/src/experimental/OneCalendar/granularities/consts.ts
@@ -1,0 +1,1 @@
+export const rangeSeparator = "→"
