+++ packages/react/src/experimental/OneCalendar/granularities/day/DayView.tsx
@@ -1,15 +1,15 @@
+import { useL10n } from "@/lib/providers/l10n"
 import { Calendar } from "@/ui/calendar"
-import { endOfWeek, startOfWeek } from "date-fns"
 import { AnimatePresence, motion } from "framer-motion"
 import {
-  DayClickEventHandler,
-  DateRange as DayPickerDateRange,
+  SelectRangeEventHandler,
+  SelectSingleEventHandler,
 } from "react-day-picker"
-import { useL10n } from "../../../lib/providers/l10n"
-import { DateRange } from "../types"
-import { getLocale } from "../utils"
+import { CalendarMode, DateRange } from "../../types"
+import { getLocale } from "../../utils"
 
-interface WeekViewProps {
+interface DayViewProps {
+  mode: CalendarMode
   selected?: Date | DateRange | null
   onSelect?: (date: Date | DateRange | null) => void
   month: Date
@@ -17,13 +17,14 @@ interface WeekViewProps {
   motionDirection?: number
 }
 
-export function WeekView({
+export function DayView({
+  mode,
   selected,
   onSelect,
   month,
   onMonthChange,
   motionDirection = 1,
-}: WeekViewProps) {
+}: DayViewProps) {
   const { locale } = useL10n()
 
   const motionVariants = {
@@ -38,28 +39,34 @@ export function WeekView({
     }),
   }
 
-  const handleDayClick: DayClickEventHandler = (day, modifiers) => {
-    if (modifiers.selected) {
-      onSelect?.(null)
-      return
-    }
-
-    const weekStart = startOfWeek(day, { weekStartsOn: 1 })
-    const weekEnd = endOfWeek(day, { weekStartsOn: 1 })
-
-    onSelect?.({
-      from: weekStart,
-      to: weekEnd,
-    })
+  if (mode === "single") {
+    return (
+      <AnimatePresence
+        mode="popLayout"
+        initial={false}
+        custom={motionDirection}
+      >
+        <motion.div
+          key={month.toISOString()}
+          variants={motionVariants}
+          custom={motionDirection}
+          initial="hidden"
+          animate="visible"
+          exit="exit"
+          transition={{ duration: 0.15, ease: [0.455, 0.03, 0.515, 0.955] }}
+        >
+          <Calendar
+            mode="single"
+            selected={selected as Date}
+            onSelect={onSelect as SelectSingleEventHandler}
+            month={month}
+            locale={getLocale(locale)}
+            weekStartsOn={1}
+          />
+        </motion.div>
+      </AnimatePresence>
+    )
   }
-
-  const selectedValue: DayPickerDateRange | undefined =
-    selected instanceof Date
-      ? {
-          from: startOfWeek(selected, { weekStartsOn: 1 }),
-          to: endOfWeek(selected, { weekStartsOn: 1 }),
-        }
-      : selected || undefined
 
   return (
     <AnimatePresence mode="popLayout" initial={false} custom={motionDirection}>
@@ -75,14 +82,12 @@ export function WeekView({
         <Calendar
           key={month.toISOString()}
           mode="range"
-          selected={selectedValue}
-          onDayClick={handleDayClick}
+          selected={selected as DateRange}
+          onSelect={onSelect as SelectRangeEventHandler}
           month={month}
           onMonthChange={onMonthChange}
           locale={getLocale(locale)}
           weekStartsOn={1}
-          showOutsideDays
-          showWeekNumber
         />
       </motion.div>
     </AnimatePresence>
