+++ packages/react/src/experimental/OneCalendar/granularities/halfyear/index.tsx
@@ -1,0 +1,98 @@
+import {
+  addMonths,
+  addYears,
+  endOfMonth,
+  startOfMonth,
+  startOfYear,
+} from "date-fns"
+import { DateRange } from "../../types"
+import { toDateRange, toDateRangeString } from "../../utils"
+import { rangeSeparator } from "../consts"
+import { GranularityDefinition } from "../types"
+import { HalfYearView } from "./HalfyearView"
+
+const formatHalfYear = (date: Date) => {
+  const month = date.getMonth()
+  const halfYear = Math.floor(month / 6)
+  return `H${halfYear + 1} ${date.getFullYear()}`
+}
+
+const toRangeString = (date: Date | DateRange | undefined | null) => {
+  const dateRange = toDateRange(date)
+  if (!dateRange) {
+    return {
+      from: "",
+      to: undefined,
+    }
+  }
+
+  const from = formatHalfYear(dateRange.from)
+  const to = dateRange.to ? formatHalfYear(dateRange.to) : undefined
+
+  return {
+    from,
+    to: to && from !== to ? to : undefined,
+  }
+}
+
+export const halfyearGranularity: GranularityDefinition = {
+  toRangeString: (date) => toRangeString(date),
+  toString: (date) => {
+    const dateRange = toRangeString(date)
+    if (!dateRange) {
+      return "-"
+    }
+    const { from, to } = dateRange
+
+    return `${from}${to && from !== to ? ` ${rangeSeparator} ${to}` : ""}`
+  },
+  fromString: (dateStr) => {
+    const dateRangeString = toDateRangeString(dateStr)
+    if (!dateRangeString) {
+      return null
+    }
+
+    const { from: fromStr, to: toStr } = dateRangeString
+
+    const parseDate = (dateStr: string) => {
+      const trimmed = dateStr.trim()
+
+      const [halfYearStr, yearStr] = trimmed.split(/\s+/)
+
+      const year = isNaN(Number(yearStr)) ? new Date().getFullYear() : +yearStr
+
+      const halfYear = Number(halfYearStr.replace(/[hH\s+]/g, "").trim())
+      return new Date(year, (halfYear - 1) * 6, 1)
+    }
+
+    return {
+      from: startOfMonth(parseDate(fromStr)),
+      to: toStr ? endOfMonth(parseDate(toStr)) : undefined,
+    }
+  },
+  navigate: (date, direction) => {
+    return addMonths(date, direction * 6)
+  },
+  navigateUIView: (viewDate, direction) => {
+    return addYears(viewDate, direction * 5)
+  },
+  label: (viewDate) => {
+    const baseYear = Math.floor(viewDate.getFullYear() / 5) * 5
+    const endYear = baseYear + 4
+    return `${baseYear} ${rangeSeparator} ${endYear}`
+  },
+  getViewDateFromDate: (date) => {
+    return startOfYear(date)
+  },
+  render: (renderProps) => {
+    return (
+      <HalfYearView
+        mode={renderProps.mode}
+        year={renderProps.viewDate.getFullYear()}
+        selected={renderProps.selected}
+        onSelect={renderProps.onSelect}
+        motionDirection={renderProps.motionDirection}
+      />
+    )
+  },
+}
