+++ packages/react/src/experimental/OneCalendar/granularities/month/MonthView.tsx
@@ -1,3 +1,4 @@
+import { useI18n } from "@/lib/providers/i18n"
 import { cn, focusRing } from "@/lib/utils"
 import {
   endOfMonth,
@@ -7,8 +8,7 @@ import {
   startOfMonth,
 } from "date-fns"
 import { AnimatePresence, motion } from "framer-motion"
-import { useI18n } from "../../../lib/providers/i18n"
-import { CalendarMode, DateRange } from "../types"
+import { CalendarMode, DateRange } from "../../types"
 
 interface MonthViewProps {
   mode: CalendarMode
